Initial renaming after forking from helloworld-confidential-client-sof
