+++ README.md
@@ -1,25 +1,8 @@
-Helloworld Confidential Client Backend for SMART-on-FHIR Frontend
+ISACC Messaging Service
 ===================
-Helloworld SMART-on-FHIR confidential client and proxy
-
-Intended to be the source project / template for forking when a confidential
-client in a SoF setting is needed.  Use in collaboration with SoF front-ends
-designed to be at least optionally run as confidential front-ends.
-
-Launch
-------
-To start an EHR launch follow the below steps
-
-Generate a launch URL in the below form, where `SERVER_NAME` matches the value in `sof_wrapper.env`
-
-http://SERVER_NAME/auth/launch
-
-Enter the above URL as the **App Launch URL** in the [SMART App Launcher](https://launch.smarthealthit.org/)
-
-* Uncheck **Simulate launch within the EHR user interface**
-* Select **R4** for **FHIR Version**
-
-Click **Launch App!**
+Backend service facilitating ISACC messaging needs.  Monitors the configured
+FHIR store for CommunicationRequests and generates Communications including
+firing SMS via contained APIs.
 
 Development
 -----------
@@ -27,7 +10,7 @@ To start the application follow the below steps in the checkout root
 
 Copy default environment variable file and modify as necessary
 
-    cp confidential_backend.env.default confidential_backend.env
+    cp messaging.env.default messaging.env
 
 Build the docker image. Should only be necessary on first run or if dependencies change.
