--- confidential_backend/api/__init__.py
@@ -1,6 +1,0 @@
-from confidential_backend.api import views, fhir
-
-__all__ = [
-    'fhir',
-    'views',
-]
