--- confidential_backend/auth/__init__.py
@@ -1,5 +1,0 @@
-from confidential_backend.auth import views
-
-__all__ = [
-    'views'
-]
