--- confidential_backend/wsgi.py
@@ -1,3 +1,0 @@
-from confidential_backend.app import create_app
-
-app = create_app()
