+++ setup.cfg
@@ -2,10 +2,10 @@
 # https://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
 
 [metadata]
-name = confidential_backend
+name = isacc_messaging
 
 [options]
-packages = confidential_backend
+packages = isacc_messaging
 zip_safe = False
 include_package_data = True
