+++ src/api/__init__.py
@@ -1,0 +1,6 @@
+from src.api import views, fhir
+
+__all__ = [
+    'fhir',
+    'views',
+]
