+++ src/auth/__init__.py
@@ -1,0 +1,5 @@
+from src.auth import views
+
+__all__ = [
+    'views'
+]
