+++ src/auth/views.py
@@ -1,9 +1,9 @@
 from flask import Blueprint, current_app, redirect, request, url_for, session
 import requests
 
-from confidential_backend.audit import audit_entry
-from confidential_backend.auth.helpers import extract_payload, format_as_jwt
-from confidential_backend.extensions import oauth
+from src.audit import audit_entry
+from src.auth.helpers import extract_payload, format_as_jwt
+from src.extensions import oauth
 
 
 # SMIT launch token encoding scheme
